\name{as}
\alias{as}
\title{construct from ranges list by coercion}
\description{
  construct from ranges list by coercion

  coerce back to IRanges

  construct from GRanges object via coercion

  coercion from GIntervalTree to GRanges object
}
\seealso{
  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{GIntervalTree-class}}, \code{\link{[}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{length}}, \code{\link{ranges}},
  \code{\link{seqinfo}}, \code{\link{seqnames}},
  \code{\link{strand}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{GIntervalTree-class}}, \code{\link{[}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{length}}, \code{\link{ranges}},
  \code{\link{seqinfo}}, \code{\link{seqnames}},
  \code{\link{strand}}

  Other IntervalTreeList:
  \code{\link{IntervalTreeList-class}},
  \code{\link{SimpleIntervalTreeList-class}}
}

