\name{findOverlaps,GRanges,GIntervalTree-method}
\alias{findOverlaps,GRanges,GIntervalTree-method}
\title{findOverlaps in an GIntervalTree}
\description{
  findOverlaps in an GIntervalTree
}
\seealso{
  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{GIntervalTree-class}}, \code{\link{[}},
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{length}},
  \code{\link{ranges}}, \code{\link{seqinfo}},
  \code{\link{seqnames}}, \code{\link{strand}}
}

