\name{findOverlaps,Ranges,PartitionedIntervalTree-method}
\alias{findOverlaps,Ranges,PartitionedIntervalTree-method}
\title{findOverlaps in a PartitionedIntervalTree}
\description{
  a partition factor must be passed
}
\seealso{
  Other PartitionedIntervalTree:
  \code{\link{PartitionedIntervalTree}},
  \code{\link{PartitionedIntervalTree}}, \code{\link{[}},
  \code{\link{end}}, \code{\link{length}},
  \code{\link{start}}
}

