\name{GIntervalTree-class}
\alias{GIntervalTree-class}
\alias{[}
\alias{length}
\alias{ranges}
\alias{seqinfo}
\alias{seqnames}
\alias{strand}
\title{GIntervalTree class}
\description{
  Defines persistent interval trees for GRanges objects.

  this walks through the Interval Trees, should be avoided

  ranges accessor

  strand accessor

  seqinfo accessor

  The GenomicRanges method uses seqnames which we should
  avoid in GIntervalTree

  subsetting
}
\seealso{
  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}}
}

