\name{GIntervalTree}
\alias{GIntervalTree}
\title{constructor function using GRanges object}
\usage{
  GIntervalTree(x)
}
\description{
  constructor function using GRanges object
}
\seealso{
  Other GIntervalTree: \code{\link{GIntervalTree-class}},
  \code{\link{[}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{length}}, \code{\link{ranges}},
  \code{\link{seqinfo}}, \code{\link{seqnames}},
  \code{\link{strand}}
}

