\name{length}
\alias{length}
\title{length accessor}
\description{
  length accessor
}
\seealso{
  Other PartitionedIntervalTree:
  \code{\link{PartitionedIntervalTree}},
  \code{\link{PartitionedIntervalTree}}, \code{\link{[}},
  \code{\link{end}},
  \code{\link{findOverlaps,Ranges,PartitionedIntervalTree-method}},
  \code{\link{start}}
}

