\name{[}
\alias{[}
\alias{end}
\alias{start}
\title{subset method}
\description{
  subset method

  start accessor

  end accessor
}
\seealso{
  Other PartitionedIntervalTree:
  \code{\link{PartitionedIntervalTree}},
  \code{\link{PartitionedIntervalTree}},
  \code{\link{findOverlaps,Ranges,PartitionedIntervalTree-method}},
  \code{\link{length}}

  Other PartitionedIntervalTree:
  \code{\link{PartitionedIntervalTree}},
  \code{\link{PartitionedIntervalTree}},
  \code{\link{findOverlaps,Ranges,PartitionedIntervalTree-method}},
  \code{\link{length}}
}

