\name{PartitionedIntervalTree}
\alias{PartitionedIntervalTree}
\title{definition}
\usage{
  PartitionedIntervalTree(ranges, partition)
}
\description{
  definition

  constructor
}
\seealso{
  Other PartitionedIntervalTree: \code{\link{[}},
  \code{\link{end}},
  \code{\link{findOverlaps,Ranges,PartitionedIntervalTree-method}},
  \code{\link{length}}, \code{\link{start}}
}

