\name{subsetByOverlaps,GIntervalTree,GRanges-method}
\alias{subsetByOverlaps,GIntervalTree,GRanges-method}
\title{subset a GIntervalTree object by overlaps
returns a GRanges object not another GIntervalTree object}
\description{
  subset a GIntervalTree object by overlaps returns a
  GRanges object not another GIntervalTree object
}
\seealso{
  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{GIntervalTree-class}}, \code{\link{as}},
  \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{intervalTrees}}, \code{\link{intervalTrees}},
  \code{\link{rangeMap}}, \code{\link{rangeMap}},
  \code{\link{ranges}}, \code{\link{seqinfo}},
  \code{\link{seqnames}}, \code{\link{strand}}
}

