\name{as}
\alias{as}
\title{construct from GRanges object via coercion}
\description{
  construct from GRanges object via coercion

  coercion from GIntervalTree to GRanges object

  construct from ranges list by coercion
}
\seealso{
  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{GIntervalTree-class}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{intervalTrees}}, \code{\link{intervalTrees}},
  \code{\link{rangeMap}}, \code{\link{rangeMap}},
  \code{\link{ranges}}, \code{\link{seqinfo}},
  \code{\link{seqnames}}, \code{\link{strand}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{GIntervalTree-class}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{intervalTrees}}, \code{\link{intervalTrees}},
  \code{\link{rangeMap}}, \code{\link{rangeMap}},
  \code{\link{ranges}}, \code{\link{seqinfo}},
  \code{\link{seqnames}}, \code{\link{strand}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}

  Other IntervalTreeList:
  \code{\link{IntervalTreeList-class}},
  \code{\link{SimpleIntervalTreeList-class}}
}

