\name{findOverlaps,GRanges,GIntervalTree-method}
\alias{findOverlaps,GRanges,GIntervalTree-method}
\title{findOverlaps in an GIntervalTree}
\description{
  findOverlaps in an GIntervalTree
}
\seealso{
  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{GIntervalTree-class}}, \code{\link{as}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{intervalTrees}}, \code{\link{intervalTrees}},
  \code{\link{rangeMap}}, \code{\link{rangeMap}},
  \code{\link{ranges}}, \code{\link{seqinfo}},
  \code{\link{seqnames}}, \code{\link{strand}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}
}

