\name{GIntervalTree-class}
\alias{GIntervalTree-class}
\alias{intervalTrees}
\alias{rangeMap}
\alias{ranges}
\alias{seqinfo}
\alias{seqnames}
\alias{strand}
\title{GIntervalTree class}
\description{
  Defines persistent interval trees for GRanges objects.

  seqnames accessor

  ranges accessor

  strand accessor

  seqinfo accessor

  intervalTrees accessor generic

  intervalTrees accessor method

  rangesMap accessor generic

  rangesMap accessor method
}
\seealso{
  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}

  Other GIntervalTree: \code{\link{GIntervalTree}},
  \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}
}

