\name{GIntervalTree}
\alias{GIntervalTree}
\title{constructor function using GRanges object}
\usage{
  GIntervalTree(x)
}
\description{
  constructor function using GRanges object
}
\seealso{
  Other GIntervalTree: \code{\link{GIntervalTree-class}},
  \code{\link{as}}, \code{\link{as}}, \code{\link{as}},
  \code{\link{findOverlaps,GRanges,GIntervalTree-method}},
  \code{\link{intervalTrees}}, \code{\link{intervalTrees}},
  \code{\link{rangeMap}}, \code{\link{rangeMap}},
  \code{\link{ranges}}, \code{\link{seqinfo}},
  \code{\link{seqnames}}, \code{\link{strand}},
  \code{\link{subsetByOverlaps,GIntervalTree,GRanges-method}}
}

